backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
djangorestframework-simplejwt>=5.3

# Authentication
django-allauth>=0.59
dj-rest-auth>=5.0

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# Task Queue
celery>=5.3
redis>=5.0

# AI
anthropic>=0.18

# PDF Generation
weasyprint>=60.1

# Utilities
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
django-ratelimit>=4.1

# Development
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
